[update] 1005 Calculator - updated divide method to return double

+++ src/java_1005/Calculator.java
@@ -18,7 +18,7 @@ public class Calculator {
         return a * b;
     }
 
-    public int divide(int a, int b) {
-        return a / b;
+    public double divide(int a, int b) {
+        return (double) a / (double) b;
     }
 }
